+++ CHANGELOG.md
@@ -5,7 +5,7 @@ By far most changes relate to `atomic-server`, so if not specified, assume the c
 **Changes to JS assets (including the front-end and JS libraries) are not shown here**, but in [`/browser/CHANGELOG`](/browser/CHANGELOG.md).
 See [STATUS.md](server/STATUS.md) to learn more about which features will remain stable.
 
-## [UNRELEASED]
+## [v0.40.0] - 2024-10-07
 
 - Speed up Commits by bundling DB transactions #297
 - Introduce `Db::apply_transaction` and `Storelike::apply_commit`